+++ src/main/java/onlyoffice/OnlyOfficeHistoryServlet.java
@@ -18,6 +18,7 @@
 
 package onlyoffice;
 
+import com.atlassian.annotations.security.UnrestrictedAccess;
 import com.atlassian.confluence.core.DateFormatter;
 import com.atlassian.confluence.core.FormatSettingsManager;
 import com.atlassian.confluence.languages.LocaleManager;
@@ -27,36 +28,26 @@ import com.atlassian.confluence.user.ConfluenceUser;
 import com.atlassian.confluence.user.ConfluenceUserPreferences;
 import com.atlassian.confluence.user.UserAccessor;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.gson.Gson;
-import com.onlyoffice.model.common.Changes;
 import com.onlyoffice.model.common.User;
 import com.onlyoffice.model.documenteditor.HistoryData;
-import com.onlyoffice.model.documenteditor.callback.History;
 import com.onlyoffice.model.documenteditor.history.Version;
-import com.onlyoffice.model.documenteditor.historydata.Previous;
 import onlyoffice.sdk.manager.security.JwtManager;
 import com.onlyoffice.manager.settings.SettingsManager;
 import onlyoffice.sdk.manager.url.UrlManager;
 import onlyoffice.managers.auth.AuthContext;
 import onlyoffice.sdk.manager.document.DocumentManager;
 import onlyoffice.utils.attachment.AttachmentUtil;
-import org.apache.log4j.LogManager;
-import org.apache.log4j.Logger;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
 import java.io.PrintWriter;
-import java.text.DateFormat;
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -64,7 +55,6 @@ import java.util.Map;
 public class OnlyOfficeHistoryServlet extends HttpServlet {
     private static final long serialVersionUID = 1L;
     private final Logger log = LogManager.getLogger("onlyoffice.OnlyOfficeHistoryServlet");
-    private static final int BUFFER_SIZE = 10240;
 
     private final LocaleManager localeManager;
     private final FormatSettingsManager formatSettingsManager;
@@ -75,6 +65,8 @@ public class OnlyOfficeHistoryServlet extends HttpServlet {
     private final UrlManager urlManager;
     private final SettingsManager settingsManager;
     private final JwtManager jwtManager;
+
+    private final ObjectMapper objectMapper = new ObjectMapper();
 
     public OnlyOfficeHistoryServlet(final LocaleManager localeManager,
                                     final FormatSettingsManager formatSettingsManager, final UserAccessor userAccessor,
@@ -93,14 +85,12 @@ public class OnlyOfficeHistoryServlet extends HttpServlet {
     }
 
     @Override
+    @UnrestrictedAccess
     public void doGet(final HttpServletRequest request, final HttpServletResponse response)
             throws ServletException, IOException {
         String type = request.getParameter("type");
         if (type != null) {
             switch (type.toLowerCase()) {
-                case "diff":
-                    getAttachmentDiff(request, response);
-                    break;
                 case "info":
                     getAttachmentHistoryInfo(request, response);
                     break;
@@ -110,55 +100,6 @@ public class OnlyOfficeHistoryServlet extends HttpServlet {
                 default:
                     response.sendError(HttpServletResponse.SC_NOT_FOUND);
                     return;
-            }
-        } else {
-            response.sendError(HttpServletResponse.SC_NOT_FOUND);
-            return;
-        }
-    }
-
-    private void getAttachmentDiff(final HttpServletRequest request, final HttpServletResponse response)
-            throws IOException {
-        if (settingsManager.isSecurityEnabled()) {
-            String securityHeader = settingsManager.getSecurityHeader();
-            String bodySecurityHeader = request.getHeader(securityHeader);
-            String authorizationPrefix = settingsManager.getSecurityPrefix();
-            String token = (bodySecurityHeader != null && bodySecurityHeader.startsWith(authorizationPrefix))
-                    ? bodySecurityHeader.substring(authorizationPrefix.length()) : bodySecurityHeader;
-
-            if (token == null || token == "") {
-                throw new SecurityException("Expected JWT");
-            }
-
-            try {
-                String payload = jwtManager.verify(token);
-            } catch (Exception e) {
-                throw new SecurityException("JWT verification failed!");
-            }
-        }
-
-        String vkey = request.getParameter("vkey");
-        String attachmentIdString = jwtManager.readHash(vkey);
-
-        if (attachmentIdString.isEmpty()) {
-            response.sendError(HttpServletResponse.SC_NOT_FOUND);
-            return;
-        }
-
-        Long attachmentId = Long.parseLong(attachmentIdString);
-        Attachment diff = attachmentUtil.getAttachmentDiff(attachmentId);
-
-        if (diff != null) {
-            InputStream inputStream = attachmentUtil.getAttachmentData(diff.getId());
-
-            response.setContentType(diff.getMediaType());
-            response.setContentLength(inputStream.available());
-
-            byte[] buffer = new byte[BUFFER_SIZE];
-
-            OutputStream output = response.getOutputStream();
-            for (int length = 0; (length = inputStream.read(buffer)) > 0;) {
-                output.write(buffer, 0, length);
             }
         } else {
             response.sendError(HttpServletResponse.SC_NOT_FOUND);
@@ -192,9 +133,7 @@ public class OnlyOfficeHistoryServlet extends HttpServlet {
         if (attachments != null) {
             ConfluenceUserPreferences preferences = userAccessor.getConfluenceUserPreferences(user);
             DateFormatter dateFormatter = preferences.getDateFormatter(formatSettingsManager, localeManager);
-            Gson gson = new Gson();
 
-            Attachment prevVersion = null;
             Collections.reverse(attachments);
             List<Version> history = new ArrayList<>();
             for (Attachment attachment : attachments) {
@@ -209,22 +148,6 @@ public class OnlyOfficeHistoryServlet extends HttpServlet {
                         )
                         .build();
 
-                Attachment changesAttachment = attachmentUtil.getAttachmentChanges(attachment.getId());
-                if (changesAttachment != null) {
-                    if (prevVersion != null && (attachment.getVersion() - prevVersion.getVersion()) == 1) {
-                        InputStream changesSteam = attachmentUtil.getAttachmentData(changesAttachment.getId());
-
-                        ObjectMapper mapper = new ObjectMapper();
-                        History changes = mapper.readValue(changesSteam, History.class);
-                        changes.setChanges(formatChanges(changes.getChanges(), user));
-
-                        version.setServerVersion(changes.getServerVersion());
-                        version.setChanges(changes.getChanges());
-                    } else {
-                        attachmentUtil.removeAttachmentChanges(attachment.getId());
-                    }
-                }
-                prevVersion = attachment;
                 history.add(version);
             }
 
@@ -235,28 +158,11 @@ public class OnlyOfficeHistoryServlet extends HttpServlet {
 
             response.setContentType("application/json");
             PrintWriter writer = response.getWriter();
-            writer.write(gson.toJson(historyInfo));
+            writer.write(objectMapper.writeValueAsString(historyInfo));
         } else {
             response.sendError(HttpServletResponse.SC_NOT_FOUND);
             return;
         }
-    }
-
-    private List<Changes> formatChanges(final List<Changes> changes, final ConfluenceUser user) {
-        ConfluenceUserPreferences preferences = userAccessor.getConfluenceUserPreferences(user);
-        DateFormatter dateFormatter = preferences.getDateFormatter(formatSettingsManager, localeManager);
-        DateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
-
-        for (Changes changesEntity : changes) {
-            try {
-                Date created = dateFormat.parse(changesEntity.getCreated());
-                changesEntity.setCreated(dateFormatter.formatDateTime(created));
-            } catch (ParseException e) {
-                log.error(e);
-            }
-        }
-
-        return changes;
     }
 
     private void getAttachmentHistoryData(final HttpServletRequest request, final HttpServletResponse response)
@@ -286,10 +192,8 @@ public class OnlyOfficeHistoryServlet extends HttpServlet {
 
         List<Attachment> attachments = attachmentUtil.getAllVersions(attachmentId);
         if (attachments != null) {
-            Gson gson = new Gson();
             HistoryData historyData = null;
 
-            Attachment prevVersion = null;
             Collections.reverse(attachments);
             for (Attachment attachment : attachments) {
                 if (attachment.getVersion() == version) {
@@ -299,28 +203,8 @@ public class OnlyOfficeHistoryServlet extends HttpServlet {
                             .url(urlManager.getFileUrl(String.valueOf(attachment.getId())))
                             .fileType(attachment.getFileExtension())
                             .build();
-
-                    Attachment diff = attachmentUtil.getAttachmentDiff(attachment.getId());
-                    if (prevVersion != null && diff != null) {
-                        boolean adjacentVersions = (attachment.getVersion() - prevVersion.getVersion()) == 1;
-                        if (adjacentVersions) {
-                            historyData.setChangesUrl(urlManager.getAttachmentDiffUri(attachment.getId()));
-                            historyData.setPrevious(Previous.builder()
-                                            .key(
-                                                    documentManager.getDocumentKey(
-                                                            String.valueOf(prevVersion.getId()),
-                                                            false
-                                                    )
-                                            )
-                                            .url(urlManager.getFileUrl(String.valueOf(prevVersion.getId())))
-                                            .fileType(prevVersion.getFileExtension())
-                                            .build()
-                            );
-                        }
-                    }
                     break;
                 }
-                prevVersion = attachment;
             }
 
             if (historyData != null) {
@@ -334,7 +218,7 @@ public class OnlyOfficeHistoryServlet extends HttpServlet {
 
                 response.setContentType("application/json");
                 PrintWriter writer = response.getWriter();
-                writer.write(gson.toJson(historyData));
+                writer.write(objectMapper.writeValueAsString(historyData));
             } else {
                 response.sendError(HttpServletResponse.SC_NOT_FOUND);
                 return;