+++ src/main/java/onlyoffice/OnlyOfficeTestServlet.java
@@ -18,10 +18,11 @@
 
 package onlyoffice;
 
+import com.atlassian.annotations.security.UnrestrictedAccess;
 import com.onlyoffice.manager.settings.SettingsManager;
 import onlyoffice.sdk.manager.security.JwtManager;
-import org.apache.log4j.LogManager;
-import org.apache.log4j.Logger;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
@@ -42,6 +43,7 @@ public class OnlyOfficeTestServlet extends HttpServlet {
     }
 
     @Override
+    @UnrestrictedAccess
     public void doGet(final HttpServletRequest request, final HttpServletResponse response)
             throws ServletException, IOException {
         if (settingsManager.isSecurityEnabled()) {